When hashing set to false, WaitForTransactionManager handles all incoming transactions
Threads register a wait for transaction and then call the startWait method which pauses the thread. When a new transaction is recieved, the wait for transaction manager compares the transaction being waited for with the current transaction; If they match, we notify the method, which allows the thread to continue.